feat: 时长支持小数输入并优化布局
- 时长字段从整数改为浮点数，支持小数输入（保留一位小数）
- 显示格式：整数显示为 1h，小数显示为 1.5h
- 调整最小窗口尺寸防止内容被遮挡
- 优化输入卡片布局宽度计算

+++ src/db.rs
@@ -7,7 +7,7 @@ pub struct Record {
     pub date: String,
     pub boss: String,
     pub income: f64,
-    pub duration: Option<i32>,   // 服务时长(小时)
+    pub duration: Option<f64>,   // 服务时长(小时)，支持小数
     pub game: Option<String>,    // 游戏名称
     pub settled: bool,           // 是否结清
 }
@@ -63,7 +63,7 @@ impl Database {
         date: &str,
         boss: &str,
         income: f64,
-        duration: Option<i32>,
+        duration: Option<f64>,
         game: Option<&str>,
         settled: bool,
     ) -> Result<()> {
